src/test/java/cz/princu/adventofcode/year21/days/Day23Test.java
package cz.princu.adventofcode.year21.days;

import cz.princu.adventofcode.common.Day;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Day23Test {

    final String testData = "" +
            "#############\n" +
            "#...........#\n" +
            "###B#C#B#D###\n" +
            "  #A#D#C#A#\n" +
            "  #########";

    Day tested = new Day23();

    @Test
    void testPart1_simpleData() {

        assertEquals(0L, tested.part1(testData));

    }

    @Test
    void testPart2_simpleData() {

        assertEquals(0L, tested.part2(testData));

    }

}